+++ core/Cargo.toml
@@ -24,7 +24,7 @@ byteorder = "1.3"
 vek = "0.9"
 nalgebra-glm = "0.6"
 lazy_static = "1.4"
-uuid = "0.8"
+uuid = { version = "0.8", features = ["serde"] }
 log = "0.4"
 num-traits = "0.2"
 num-derive = "0.3"
